Password Based Key Derivation is now based on an OutputFeedbackRandomByteSequence with multiple feedback-iterations (>1000)
AES-with convenience Key-Generation... 
Is that ok?
Should it be in a higher Class, so OneTimePad can also be initialized with a password?

Tests provide proof that two different instances of AES provide the 'same' keying material.
Tests provide proof that two different instances of AES are capable of en- and decript correctly provided the same password.